Refactor WebGPU renderer into stage modules
